# feat: add dp attention support for Qwen 2/3 MoE models, fixes #6088