--- a/python/sglang/srt/models/qwen3_moe.py
+++ b/python/sglang/srt/models/qwen3_moe.py
@@ -184,6 +184,7 @@ class Qwen3MoeAttention(nn.Module):
             quant_config=quant_config,
             tp_rank=attn_tp_rank,
             tp_size=attn_tp_size,
+            reduce_results=not global_server_args_dict["enable_dp_attention"],
             prefix=add_prefix("o_proj", prefix),
         )