This is the prerequisites of EP, which is introduced in PR #5917

<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

As described in #6088, DP attention is not supported for Qwen MoE models, but #5917 introduces EP MoE for them. This PR introduces DP attention support for them.

## Modifications

<!-- Describe the changes made in this PR. -->

Following DP Attention part in `deepseek_v2.py`, I modified `qwen2_moe.py` and `qwen3_moe.py`.

## Benchmark Results and Accuracy Results

> The following benchmark results contains in-negligible measurement error, see https://github.com/sgl-project/sglang/pull/6121#issuecomment-2869937995

Using 4xA40 (PCIe).

Baseline: TP=4 before this PR (commit f1ff736d681414dd97316ef4fb66378287c3d21e).

Commands:

```bash
# bench_one_batch
python -m sglang.bench_one_batch \
    --model-path Qwen/Qwen3-30B-A3B \
    --mem-fraction-static 0.6 \
    --batch 64 \
    --input-len 256 \
    --output-len 32 \
    --tp-size 4 ...

# bench_offline_throughput
python -m sglang.bench_offline_throughput \
    --model-path Qwen/Qwen3-30B-A3B \
    --mem-fraction-static 0.6 \
    --num-prompts 16 \
    --tp-size 4 ...

# bench_serving server
python -m sglang.launch_server \
    --model-path Qwen/Qwen3-30B-A3B \
    --trust-remote-code \
    --mem-fraction-static 0.6 \
    --max-prefill-tokens 2048 \
    --chunked-prefill-size 2048 \
    --disable-radix-attention \
    --tp-size 4

# bench_serving client
python3 -m sglang.bench_serving \
    --backend sglang \
    --dataset-name random \
    --dataset-path ShareGPT.json \
    --model Qwen3-30B-A3B \
    --host 127.0.0.1 \
    --random-range-ratio 1.0 \
    --num-prompts 512 \
    --random-input-len 512 \
    --random-output-len 256 \
    --max-concurrency 128
```

- Benchmark Results(Qwen3-30B-A3B):
  - `sglang.bench_one_batch`

    |     Configuration    | Prefill Throughput | Decode Throughput |
    |----------------------|--------------------|-------------------|
    |   TP = 4 (baseline)  |      17397.76      |      2628.55      |
    | TP = 4, No DP, No EP |      17677.90      |      2487.72      |
    |  TP = DP = 4, No EP  |       4561.55      |      600.84       |
    |   TP = DP = EP = 4   |       2114.99      |      649.72       |

  - `sglang.bench_offline_throughput`

    |     Configuration    | Duration (s) | Total token Throughput |
    |----------------------|--------------|------------------------|
    |   TP = 4 (baseline)  |    13.60     |         833.01         |
    | TP = 4, No DP, No EP |    13.85     |         817.86         |
    |  TP = DP = 4, No EP  |    27.97     |         405.01         |
    |   TP = DP = EP = 4   |    31.49     |         359.66         |
  
  - `sglang.bench_serving`

    |     Configuration    | Duration (s) | Median TTFT (ms) | Median ITL (ms) |
    |----------------------|--------------|------------------|-----------------|
    |   TP = 4 (baseline)  |    64.81     |      2678.38     |      45.51      |
    | TP = 4, No DP, No EP |    68.40     |      2979.11     |      45.00      |
    |  TP = DP = 4, No EP  |    75.51     |      3188.26     |      48.65      |
    |   TP = DP = EP = 4   |    93.75     |      4617.38     |      55.41.     |

    <details>
    <summary>Detailed results</summary>

    - TP = 4 (baseline)
      ```log
      ============ Serving Benchmark Result ============
      Backend:                                 sglang
      Traffic request rate:                    inf
      Max reqeuest concurrency:                128
      Successful requests:                     512
      Benchmark duration (s):                  64.81
      Total input tokens:                      262144
      Total generated tokens:                  131072
      Total generated tokens (retokenized):    131071
      Request throughput (req/s):              7.90
      Input token throughput (tok/s):          4044.97
      Output token throughput (tok/s):         2022.48
      Total token throughput (tok/s):          6067.45
      Concurrency:                             127.54
      ----------------End-to-End Latency----------------
      Mean E2E Latency (ms):                   16143.09
      Median E2E Latency (ms):                 16095.80
      ---------------Time to First Token----------------
      Mean TTFT (ms):                          2667.26
      Median TTFT (ms):                        2678.38
      P99 TTFT (ms):                           4914.29
      ---------------Inter-Token Latency----------------
      Mean ITL (ms):                           52.85
      Median ITL (ms):                         45.51
      P95 ITL (ms):                            55.12
      P99 ITL (ms):                            62.50
      Max ITL (ms):                            4268.96
      ==================================================
      ```
    - TP = 4, No DP, No EP
      ```log
      ============ Serving Benchmark Result ============
      Backend:                                 sglang
      Traffic request rate:                    inf
      Max reqeuest concurrency:                128
      Successful requests:                     512
      Benchmark duration (s):                  68.40
      Total input tokens:                      262144
      Total generated tokens:                  131072
      Total generated tokens (retokenized):    131071
      Request throughput (req/s):              7.48
      Input token throughput (tok/s):          3832.30
      Output token throughput (tok/s):         1916.15
      Total token throughput (tok/s):          5748.46
      Concurrency:                             127.54
      ----------------End-to-End Latency----------------
      Mean E2E Latency (ms):                   17038.84
      Median E2E Latency (ms):                 16078.40
      ---------------Time to First Token----------------
      Mean TTFT (ms):                          2994.61
      Median TTFT (ms):                        2979.11
      P99 TTFT (ms):                           6363.70
      ---------------Inter-Token Latency----------------
      Mean ITL (ms):                           55.08
      Median ITL (ms):                         45.00
      P95 ITL (ms):                            55.40
      P99 ITL (ms):                            61.72
      Max ITL (ms):                            7124.37
      ==================================================
      ```
    - TP = DP = 4, No EP
      ```log
      ============ Serving Benchmark Result ============
      Backend:                                 sglang
      Traffic request rate:                    inf
      Max reqeuest concurrency:                128
      Successful requests:                     512
      Benchmark duration (s):                  75.51
      Total input tokens:                      262144
      Total generated tokens:                  131072
      Total generated tokens (retokenized):    131070
      Request throughput (req/s):              6.78
      Input token throughput (tok/s):          3471.75
      Output token throughput (tok/s):         1735.88
      Total token throughput (tok/s):          5207.63
      Concurrency:                             127.70
      ----------------End-to-End Latency----------------
      Mean E2E Latency (ms):                   18833.12
      Median E2E Latency (ms):                 17459.89
      ---------------Time to First Token----------------
      Mean TTFT (ms):                          3275.09
      Median TTFT (ms):                        3188.26
      P99 TTFT (ms):                           8554.38
      ---------------Inter-Token Latency----------------
      Mean ITL (ms):                           61.01
      Median ITL (ms):                         48.65
      P95 ITL (ms):                            58.50
      P99 ITL (ms):                            64.15
      Max ITL (ms):                            8558.96
      ==================================================
      ```
    - TP = DP = EP = 4
      ```log
      ============ Serving Benchmark Result ============
      Backend:                                 sglang
      Traffic request rate:                    inf
      Max reqeuest concurrency:                128
      Successful requests:                     512
      Benchmark duration (s):                  93.75
      Total input tokens:                      262144
      Total generated tokens:                  131072
      Total generated tokens (retokenized):    131071
      Request throughput (req/s):              5.46
      Input token throughput (tok/s):          2796.26
      Output token throughput (tok/s):         1398.13
      Total token throughput (tok/s):          4194.39
      Concurrency:                             127.70
      ----------------End-to-End Latency----------------
      Mean E2E Latency (ms):                   23382.37
      Median E2E Latency (ms):                 22447.24
      ---------------Time to First Token----------------
      Mean TTFT (ms):                          4567.59
      Median TTFT (ms):                        4617.38
      P99 TTFT (ms):                           8424.25
      ---------------Inter-Token Latency----------------
      Mean ITL (ms):                           73.78
      Median ITL (ms):                         55.41
      P95 ITL (ms):                            64.97
      P99 ITL (ms):                            70.56
      Max ITL (ms):                            11509.58
      ==================================================
      ```
    </details>
- Accuracy Results (MMLU):
  - Qwen2-57B-A13B

    |     Configuration    | Accuracy |
    |----------------------|----------|
    |   TP = 4 (baseline)  |  0.728   |
    | TP = 4, No DP, No EP |  0.730   |
    |  TP = DP = 4, No EP  |  0.730   |
    |   TP = DP = EP = 4   |  0.733   |

  - Qwen3-30B-A3B

    |     Configuration    | Accuracy |
    |----------------------|----------|
    |   TP = 4 (baseline)  |  0.798   |
    | TP = 4, No DP, No EP |  0.797   |
    |  TP = DP = 4, No EP  |  0.797   |
    |   TP = DP = EP = 4   |  0.796   |


## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.