@yhyang201 2025-05-09
Does it seem like the performance might have decreased?