@yizhang2077 2025-05-11
I think `TP baseline ` and `TP = 4, No DP, No EP` are the same setting? Why the decode throughput declines ?