@Fr4nk1inCs 2025-05-11
I would investigate the performance issue soon.