@Fr4nk1inCs 2025-05-11
> I think `TP baseline ` and `TP = 4, No DP, No EP` are the same setting? Why the decode throughput declines ?

This might be due to measurement error. I've re-run `bench_one_batch` with `batch size = 128` (repeated 8 times), `input len = 512`, `output len = 128` for `TP = 4`, here's the result.

- Prefill Throughput (tokens/s)
  | Before/After this PR | Min | Median | Avg | Max |
  |----------------------|-----|--------|-----|-----|
  | Before | 17680.70 | 17798.875 | 17796.655 | 17879.61 |
  | After  | 17739.24 | 17811.515 | 17811.445 | 17865.12 |
- Decode Throughput (tokens/s)
  | Before/After this PR | Min | Median | Avg | Max |
  |----------------------|-----|--------|-----|-----|
  | Before | 3405.03 | 3504.10 | 3489.61 | 3531.65 |
  | After  | 3392.91 | 3491.015 | 3498.1425 | 3627.14 |

<details>
<summary>Full result</summary>

| Run |  Prefill (Before) | Prefill (After) | Decode (Before) | Decode (After) |
|-----|-------------------|-----------------|-----------------|-----------------|
|  1  | 17879.61 | 17865.12 | 3531.65 | 3523.08 |
|  2  | 17855.86 | 17855.13 | 3405.03 | 3515.60 |
|  3  | 17746.00 | 17839.14 | 3511.97 | 3392.91 |
|  4  | 17818.82 | 17819.87 | 3430.00 | 3466.43 |
|  5  | 17795.70 | 17739.24 | 3496.23 | 3627.14 |
|  6  | 17794.50 | 17803.16 | 3531.02 | 3431.58 |
|  7  | 17802.05 | 17784.95 | 3522.43 | 3432.50 |
|  8  | 17680.70 | 17784.95 | 3488.55 | 3595.90 |
</details>