@Fr4nk1inCs 2025-05-11
Re-run the benchmark results with repetitions:

- Benchmark Results(Qwen3-30B-A3B):
  - `sglang.bench_one_batch`

    |     Configuration    | Prefill Throughput | Decode Throughput |
    |----------------------|--------------------|-------------------|
    |   TP = 4 (baseline)  |      17397.76      |      2628.55      |
    | TP = 4, No DP, No EP |      17677.90      |      2487.72      |
    |  TP = DP = 4, No EP  |       4561.55      |      600.84       |
    |   TP = DP = EP = 4   |       2114.99      |      649.72       |

  - `sglang.bench_offline_throughput`

    |     Configuration    | Duration (s) | Total token Throughput |
    |----------------------|--------------|------------------------|
    |   TP = 4 (baseline)  |    13.60     |         833.01         |
    | TP = 4, No DP, No EP |    13.85     |         817.86         |
    |  TP = DP = 4, No EP  |    27.97     |         405.01         |
    |   TP = DP = EP = 4   |    31.49     |         359.66         |
  
  - `sglang.bench_serving`

    |     Configuration    | Duration (s) | Median TTFT (ms) | Median ITL (ms) |
    |----------------------|--------------|------------------|-----------------|
    |   TP = 4 (baseline)  |    64.81     |      2678.38     |      45.51      |
    | TP = 4, No DP, No EP |    68.40     |      2979.11     |      45.00      |
    |  TP = DP = 4, No EP  |    75.51     |      3188.26     |      48.65      |
    |   TP = DP = EP = 4   |    93.75     |      4617.38     |      55.41.     |