@Fr4nk1inCs 2025-05-11
Re-run the benchmarks with repetitions (repeated 8 times, results represented as `avg ± std`):

- `sglang.bench_one_batch`

  |     Configuration    | Prefill Throughput | Decode Throughput |
  |----------------------|--------------------|-------------------|
  |   TP = 4 (baseline)  | 17778.80 ± 36.54 | 2530.26 ± 60.64 |
  | TP = 4, No DP, No EP | 17775.3825 ± 49.82 | 2555.96 ± 97.14 |
  |  TP = DP = 4, No EP  | 4561.97 ± 3.83 | 602.78 ± 73.29 |
  |   TP = DP = EP = 4   | 2115.43 ± 1.51 | 594.87 ± 68.65 |

- `sglang.bench_offline_throughput` **(radix cache disabled, one warm up run)**

  |     Configuration    | Duration (s) | Total token Throughput |
  |----------------------|--------------|------------------------|
  |   TP = 4 (baseline)  | 9.42 ± 0.05 | 1202.09 ± 6.78 |
  | TP = 4, No DP, No EP | 9.38 ± 0.07 | 1206.99 ± 8.39 |
  |  TP = DP = 4, No EP  | 11.02 ± 0.28 | 1028.37 ± 24.32 |
  |   TP = DP = EP = 4   | 15.48 ± 0.24 | 731.94 ± 10.99 |

- `sglang.bench_serving` (two warm up run, benched only 4 times to save time)

  |     Configuration    | Duration (s) | Median TTFT (ms) | Median ITL (ms) |
  |----------------------|--------------|------------------|-----------------|
  |   TP = 4 (baseline)  | 64.72 ± 0.36 | 2563.87 ± 72.88 | 45.10 ± 0.09 |
  | TP = 4, No DP, No EP | 65.00 ± 0.32 | 2599.49 ± 46.41 | 45.08 ± 0.10 |
  |  TP = DP = 4, No EP  | 68.02 ± 0.22 | 2406.02 ± 51.81 | 48.66 ± 0.08 |
  |   TP = DP = EP = 4   | 88.76 ± 0.08 | 4243.19 ± 38.45 | 55.47 ± 0.03 |

At least this PR didn't introduce performance degradation for TP = 4 😇.